+++ src/th06_102h.exe/Ending/OnDraw.c
@@ -10,8 +10,8 @@ undefined4 Ending::OnDraw(Ending *ending)
   
   rectHeight = 480;
   rectWidth = 640;
-  subFrame = __ftol2((double)(ending->anmTimer4).subFrame);
-  prevFrame = __ftol2((double)(float)(ending->anmTimer4).previous);
+  subFrame = __ftol2((ending->anmTimer4).subFrame);
+  prevFrame = __ftol2((float)(ending->anmTimer4).previous);
   AnmManager::drawEndingRect(g_AnmManager,0,0,0,prevFrame,subFrame,rectWidth,rectHeight);
   for (i = 0; i < 16; i = i + 1) {
     if ((&ending->AnmVm)[i].anmFileIndex != 0) {
