+++ src/th06_102h.exe/Global/FUN_0042fc80.c
@@ -6,8 +6,8 @@ bool FUN_0042fc80(int param_1)
   long lVar2;
   
   if (*(int *)(param_1 + 0x14) != 0) {
-    lVar2 = __ftol2((double)((((float)*(int *)(param_1 + 0x2c) + *(float *)(param_1 + 0x28)) * 255.0
-                             ) / (float)*(int *)(param_1 + 0x14)));
+    lVar2 = __ftol2((((float)*(int *)(param_1 + 0x2c) + *(float *)(param_1 + 0x28)) * 255.0) /
+                    (float)*(int *)(param_1 + 0x14));
     *(long *)(param_1 + 0x10) = lVar2;
     if (*(int *)(param_1 + 0x10) < 0) {
       *(undefined4 *)(param_1 + 0x10) = 0;
